Raw instance data.
