+++ odc/examples/camera.rs
@@ -1,12 +1,13 @@
 mod common;
 
 use glam::Mat4;
-use odc::{InstanceInfo, OdcCore, RenderInfo, StaticMesh, WindowSize};
+use odc::{OdcCore, RenderInfo, StaticMesh, WindowSize};
 use std::f32::consts::PI;
 use std::time::Instant;
 use vp_cam::{Camera, CameraBuilder, Vec3};
 use winit::event::{Event, WindowEvent};
 use winit::event_loop::{ControlFlow, EventLoop};
+use crate::common::InstanceInfo;
 
 fn main() {
     env_logger::init();
@@ -42,7 +43,7 @@ fn main() {
                 let instance = InstanceInfo {
                     transform: glam::Mat4::IDENTITY.to_cols_array_2d(),
                 };
-                renderer.write_instances(&[instance], 0);
+                renderer.write_instances(bytemuck::cast_slice(&[instance]), 0);
 
                 let draw = StaticMesh {
                     indices: 0..3,
